requirements_service.txt
requests==2.28.1
boto3==1.24.68
streaming-form-data==1.11.0
textblob==0.17.1
openai
anthropic
google-generativeai
replicate
nanoid